refactor(users): :fire: avoid duplicates
