+++ src/modules/users/users.repository.ts
@@ -98,177 +98,4 @@ export class UsersRepository extends AbstractRepository<User> {
 
         return userObjects.map((userObject) => new User(userObject));
     }
-
-    changeRole({
-        targetUser,
-        actor,
-        role,
-    }: {
-        targetUser: User;
-        actor: User;
-        role: UserRole;
-    }): User {
-        if (targetUser._id === actor._id) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        role: 'cannotChangeRoleOfYourself',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        if (actor.role !== UserRole.Manager || targetUser.role === UserRole.Manager) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        role: 'notAllowedToChangeRole',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        if (targetUser?.role === role) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        role: 'userAlreadyHasRole',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        targetUser.role = role;
-        return targetUser;
-    }
-
-    blockUser({
-        targetUser,
-        actor,
-        block: blockObject,
-    }: {
-        targetUser: User;
-        actor: User;
-        block: UpdateUserMntDto['block'];
-    }): User {
-        if (targetUser._id === actor._id) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        block: 'cannotBlockYourself',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        if (actor.role !== UserRole.Manager || targetUser.role === UserRole.Manager) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        block: 'notAllowedToBlock',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        if (targetUser?.block?.isBlocked) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        block: 'userAlreadyBlocked',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        const actLogs = (targetUser.block && targetUser?.block?.activityLogs) || [];
-        actLogs.push({
-            action: BlockAction.Block,
-            actionAt: new Date(),
-            actionBy: actor._id,
-            reason: blockObject?.activityLogs.reason ?? '',
-            note: blockObject?.activityLogs.note ?? '',
-        });
-
-        return Object.assign(targetUser, {
-            block: {
-                isBlocked: true,
-                activityLogs: actLogs,
-            },
-        });
-    }
-
-    unblockUser({
-        targetUser,
-        actor,
-        block: blockObject,
-    }: {
-        targetUser: User;
-        actor: User;
-        block: UpdateUserMntDto['block'];
-    }): User {
-        if (targetUser._id === actor._id) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        block: 'cannotUnblockYourself',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        if (actor.role !== UserRole.Manager || targetUser.role === UserRole.Manager) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        block: 'notAllowedToUnblock',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        if (targetUser?.block?.isBlocked) {
-            throw new HttpException(
-                {
-                    status: HttpStatus.UNPROCESSABLE_ENTITY,
-                    errors: {
-                        block: 'userAlreadyNotBlocked',
-                    },
-                },
-                HttpStatus.UNPROCESSABLE_ENTITY,
-            );
-        }
-
-        const actLogs = (targetUser.block && targetUser?.block?.activityLogs) || [];
-        actLogs.push({
-            action: BlockAction.Unblock,
-            actionAt: new Date(),
-            actionBy: actor._id,
-            reason: blockObject?.activityLogs.reason ?? '',
-            note: blockObject?.activityLogs.note ?? '',
-        });
-
-        return Object.assign(targetUser, {
-            block: {
-                isBlocked: false,
-                activityLogs: actLogs,
-            },
-        });
-    }
 }
